requirements.txt
pbr>=0.5.21,<1.0
Babel>=1.3
six>=1.7.0
oslo.config>=1.4.0  # Apache-2.0
oslo.log>=0.1.0  # Apache-2.0
oslo.serialization>=1.0.0  # Apache-2.0
oslo.utils>=1.0.0  # Apache-2.0
taskflow>=0.6